Fix .net sdk 5.0.200 and up cracking error
The F# targets for several versions have set some properties based on the presence of the DOTNET_HOST_PATH environment variable, which the dotnet CLI sets when `dotnet build` or `dotnet msbuild` are run.  In 5.0.200 and up, some internal default values for these properties changed that resulted in failed cracking in this library due to invalid values, which brought to light that we were never setting this environment variable. As soon as we do (using the algorithm from the dotnet CLI itself to discover the path), all of the cracking issues go away.